support manipulating breakpoints while the program is running (and not stopped)
this was a rather delicate thing to do:
To manipulate breakpoints, the program must be stopped. To do so while it's running,
we keep track of the locations where we want to set/delete them, and politely ask gdb
to interrupt the execution. Once we receive the notification that it has done so,
we actually set/delete the breakpoints according to what we remembered.
If there were some pending breakpoint operations, that means the interruption was
artificially done by us, meaning that we want to resume the execution once we're done.

The difficulty was compounded by obscure "features" of gdb, such as the mi-async mode,
and the fact that "continue" and "-exec-continue" are NOT actually the same thing for gdb-mi

fixes #2